day10 part2 120x speedup
So basically we did 2 things :
1. 1 we precomputed combinations of buttons and applied them all at once,
which shaved about 40% of runtime.
2. We stopped using bitvec and used a simple wrapped usize, which ran
  100x faster !!

Technically we also do check the validity of the sum pattern istead of
applying the buttons one by one too in 2., but there is no way this saves
that much time (MAYBE 4x)

For some reason I didn't really dig into my usage of bitvec<usize> lead
to a bunch of allocs (??)
There were also a bunch of safety things which we do not care about at
all.

Anyway for our extremely simple use case we can simply keep a usise
storing the resulting bit parity pattern, and check the pattern match
with a simple equality check.