day10 part 1: enumerate + SimpleBitVec
Move to use SimpleBitVec instead of bitvec and change the algo
to use the same logic as part2 : generate the combination and test them
keeping the shortest.

I don't know if this is actually faster than pushing the buttons in my
DP way to be honest.
Probably since we don't even need to allocate the HashSet anymore.