day 11 4xfaster
With this encoding and everything statically allocated, we're about 4x
faster to ~200microsec for part2.

We could probably make the string bit about as fast by removing a few
copies with a few lifetime annotations.

This is not really computationally intensive so the wins from the
encoding are not THAT large (still 4x is pretty good)